+++ CHANGELOG.md
@@ -7,8 +7,10 @@ Please add changes to "master", preferably ordered by their significance. (Most 
 ### master
 
  - Visualization settings UI now matches the new attribute UI.
+ - When you try to overwrite a table or view, you now get a confirmation prompt instead of an error.
  - Tab-completion for LynxKite operations works in Python command-line
    and in Jupyter notebooks.
+ - You can now filter for position (Double, Double) attributes.
  - Added _Segment by geographical proximity_ operation to segment vertices using Shapefiles and
    coordinate vertex attributes. Can be used e.g. to segment by geo regions or points of interests.
 
@@ -37,6 +39,7 @@ Please add changes to "master", preferably ordered by their significance. (Most 
  - Spark Monitor doesn't send alarm in case of long-running tasks.
 
 Ecosystem:
+
  - Replace `run_task.sh` and `show_target.sh` with a new unified `tasks.sh`.
  - Renamed from "Lynx Data Automation Framework" to "Lynx Enterprise".